fix(review): address Copilot audit on PR #201 — 8 bugs + 5 docs tightens + 3 consistency gaps
Responds to 18 of the 19 unresolved Copilot review comments on PR #201.
The 19th — the 465-line module-size note on adapters/weaver_contracts.py
— is intentionally retained per the PR description's documented Mode B
choice and the existing precedent of adapters/mcp.py at 352 lines.

Bug fixes (8):
* routing/router.py — RouteResult.to_routing_decision() no longer drops
  router diagnostics when the caller supplies a metadata["contextweaver"]
  dict; merge into the existing key instead of setdefault.
* routing/router.py — to_routing_decision() now passes max_choices=
  len(self.candidate_items) so make_choice_cards' default cap of 20 does
  not silently truncate routes configured with top_k > 20.
* envelope.py — RoutingDecision.from_dict() normalises the RFC 3339 "Z"
  suffix to "+00:00" before datetime.fromisoformat, so spec-shaped
  date-time payloads parse on Python 3.10 (stdlib only learned to accept
  "Z" in 3.11).
* adapters/weaver_contracts.py — to_weaver_selectable_item() now raises
  CatalogError when the caller supplies metadata["_contextweaver"]
  instead of silently overwriting the user's reserved key.
* adapters/weaver_contracts.py — to_weaver_routing_decision() remaps
  selected_card_id (which refers to a CW 1:1 card) to the synthetic
  spec menu id when the underlying item is present; from_weaver_routing
  _decision() reverses the remap so the round-trip is lossless.
* adapters/weaver_contracts.py — from_weaver_frame() only reverses the
  "(no summary)" sentinel when metadata["_contextweaver"] proves the
  Frame came from to_weaver_frame. Foreign-origin frames that
  legitimately use that string no longer lose their summary.
* tests/test_adapters_weaver.py (new) — splits the weaver-spec test
  block out of test_adapters.py so the module-level
  pytest.importorskip("weaver_contracts") no longer suppresses MCP /
  A2A / FastMCP coverage in minimal environments.
* pyproject.toml — jsonschema pin raised to >=4.18 (the version that
  bundles `referencing`, which the conformance script's local $ref
  resolver depends on).

Documentation accuracy (5):
* envelope.py — RoutingDecision docstring spells out that to_dict()
  produces CW-shaped JSON, not spec-shaped; schema-valid output goes
  through to_weaver_routing_decision().
* AGENTS.md — same clarification on the Key Types row.
* CHANGELOG.md — same clarification on the release-note bullet.
* README.md §5 — I-05 status downgraded from "satisfied via firewall"
  to "satisfied on the Frame adapter path"; explicitly names the
  legacy raw-output ingestion APIs as non-canonical for spec compliance.
* scripts/weaver_spec_conformance.py — _check_schemas docstring spells
  out that only adapter output is validated against the spec schema.

Consistency (3):
* Makefile / AGENTS.md / docs/agent-context/workflows.md — the new
  `weaver-conformance` target is now listed in the authoritative
  command tables.
* pyproject.toml — [all] extra now includes contextweaver[weaver-spec].
* README.md §3 + §5 / docs/weaver_spec_mapping.md — clarify that the
  CI gate fetches schemas from raw.githubusercontent.com (the source
  the gate actually uses) and that weaver-spec.dev publishes the same
  documents.

Cleanup:
* scripts/weaver_spec_conformance.py — dropped the unused base_uri
  computation.

Tests: 801 pass (was 797), 1 skipped. Coverage 87%.
make ci: ruff format + ruff check + mypy strict + pytest + examples +
demo all green. weaver-conformance: 5/5 checks pass (4 round-trips + 1
JSON-Schema validation against the published spec schemas).
